Adding Navigation buttons
Page is feeling a bit cluttered with the amount of buttons now. so will try and find a way to hide some / move them around.

Especially on number main, maybe need to force the blank space inbetween the top and bottom row of buttons, at least until the game starts.