More Button working
Number page works as intended, layout might want editing. but home and next buttons only appear when the game is done as intended.

need to figure out how to have info button always render at the top and then only have the nav buttons render if the game is over